% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregateFeature}
\alias{aggregateFeature}
\alias{aggregateFeature,ComparisonData-method}
\title{Aggregate Feature Data for Comparison}
\usage{
aggregateFeature(comparison, ...)

\S4method{aggregateFeature}{ComparisonData}(comparison, upstream_bp = 1000)
}
\arguments{
\item{comparison}{A ComparisonData object}

\item{...}{Additional arguments (not used)}

\item{upstream_bp}{Number of base pairs upstream for aggregation (default is 1000)}
}
\value{
A ComparisonData object with aggregated feature data
}
\description{
Aggregate Feature Data for Comparison
}
\section{Methods (by class)}{
\itemize{
\item \code{aggregateFeature(ComparisonData)}: Method to aggregate feature data

}}
