% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{createFormattedInput}
\alias{createFormattedInput}
\alias{createFormattedInput,data.frame-method}
\title{Create Formatted Input}
\usage{
createFormattedInput(results, style)

\S4method{createFormattedInput}{data.frame}(results, style)
}
\arguments{
\item{results}{A data frame containing the comparison results}

\item{style}{The style of the formatted input (Bracket-Spacing, Letter-Digit, Simple-List, Verbose)}
}
\value{
A formatted input string
}
\description{
Create Formatted Input
}
\section{Methods (by class)}{
\itemize{
\item \code{createFormattedInput(data.frame)}: Method to create formatted input

}}
