% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{GRangesData-class}
\alias{GRangesData-class}
\title{GRangesData Class}
\description{
This class represents a GRanges object along with its metadata.
}
\section{Slots}{

\describe{
\item{\code{granges}}{A GRanges object containing genomic ranges.}

\item{\code{metadata}}{A data frame containing metadata for the GRanges object.}
}}

