% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{LLMFormatData-class}
\alias{LLMFormatData-class}
\title{LLMFormatData Class}
\description{
This class represents the LLM formatted data, including training and test datasets.
}
\section{Slots}{

\describe{
\item{\code{training_data}}{JSON formatted training data.}

\item{\code{test_data}}{JSON formatted test data.}
}}

