% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{readFile}
\alias{readFile}
\alias{readFile,character-method}
\title{Read a BED file and create a GRangesData object}
\usage{
readFile(file, ...)

\S4method{readFile}{character}(file, format = "BED")
}
\arguments{
\item{file}{Path to the BED file}

\item{...}{Additional arguments (not used)}

\item{format}{Format of the file (default is "BED")}
}
\value{
A GRangesData object
}
\description{
Read a BED file and create a GRangesData object
}
\section{Methods (by class)}{
\itemize{
\item \code{readFile(character)}: Method to read BED files

}}
