% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{toLLMFormat}
\alias{toLLMFormat}
\alias{toLLMFormat,ComparisonData-method}
\title{Convert Comparison Data to JSON Format}
\usage{
toLLMFormat(
  comparison,
  split_ratio = NULL,
  style = "Bracket-Spacing",
  system_prompt = "Analyze the following sequences:"
)

\S4method{toLLMFormat}{ComparisonData}(
  comparison,
  split_ratio = NULL,
  style = "Bracket-Spacing",
  system_prompt = "Analyze the following sequences:"
)
}
\arguments{
\item{comparison}{A ComparisonData object}

\item{split_ratio}{Ratio for splitting data into training and test sets (default is NULL)}

\item{style}{The style of the output format (Bracket-Spacing, Bracket-Comma, ABC, DEF)}

\item{system_prompt}{A character string for the system prompt in the LLM JSON output (default is "Analyze the following sequences:")}
}
\value{
A list containing JSON formatted data for training and test sets or a single JSON object
}
\description{
Convert Comparison Data to JSON Format
}
\section{Methods (by class)}{
\itemize{
\item \code{toLLMFormat(ComparisonData)}: Method to convert comparison data to JSON format

}}
